Add files via upload
A login form is one of the most important pages of a website that allows authorized users to access an entire site or a part of a website. The login/sign-up page is the first page to show the users for login-protected sites. Whether it’s a signup or login page, it should be attractive, user-friendly, and easy to use.